# frontend/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ArtifactFlow API",
    "description": "Multi-agent system API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Send Message",
        "description": "\u53d1\u9001\u65b0\u6d88\u606f\uff08multipart/form-data\uff09\n\n`payload` \u4e3a ChatRequest \u7684 JSON \u5b57\u7b26\u4e32\uff1b`files` \u4e3a\u53ef\u9009\u9644\u4ef6\u3002\u9644\u4ef6\u5728\u8d77 turn\n\u524d\u540c\u6b65\u8f6c\u6210 artifact\uff08source=user_upload\uff09\u843d\u5e93\uff0c\u5e76\u628a (id, filename) \u900f\u4f20\u8fdb\nUSER_INPUT \u4e8b\u4ef6\u6b63\u6587\uff0c\u8ba9 agent \u77e5\u9053\u54ea\u4e9b\u662f\u672c\u8f6e\u65b0\u4f20\u7684\u3002\u8fd4\u56de stream_url \u4f9b\u524d\u7aef\u8ba2\u9605\u3002",
        "operationId": "send_message_api_v1_chat_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_send_message_api_v1_chat_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "List Conversations",
        "description": "\u5217\u51fa\u5bf9\u8bdd\u5217\u8868",
        "operationId": "list_conversations_api_v1_chat_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/skills": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Admin List Skills",
        "description": "\u5217\u51fa\u6240\u6709 shared skill,\u4e0d\u6309 admin \u81ea\u5df1\u7684\u90e8\u95e8\u53ef\u89c1\u6027\u8fc7\u6ee4\u3002",
        "operationId": "admin_list_skills_api_v1_admin_skills_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSkillListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/departments": {
      "get": {
        "tags": [
          "departments"
        ],
        "summary": "List Departments",
        "description": "\u5217\u51fa\u67d0\u7236\u4e0b\u7684\u5b50\u90e8\u95e8\uff08\u6309 name \u6392\u5e8f\uff09\u3002\n\nparent_id \u7f3a\u7701 \u2192 \u4e00\u7ea7\u90e8\u95e8\uff1b\u4f20\u5177\u4f53 id \u2192 \u8be5\u90e8\u95e8\u7684\u76f4\u63a5\u5b50\u3002\n\u7ed9\u524d\u7aef cascader \u6bcf\u7ea7\u6e32\u67d3\u7528\u3002",
        "operationId": "list_departments_api_v1_departments_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "parent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Parent id; omit for top-level",
              "title": "Parent Id"
            },
            "description": "Parent id; omit for top-level"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "departments"
        ],
        "summary": "Create Department",
        "description": "\u663e\u5f0f\u521b\u5efa\u90e8\u95e8\uff08admin \u5728\u7ba1\u7406 UI \u4e2d\u70b9 \"+ \u65b0\u5efa\" \u8d70\u8fd9\u91cc\uff09\u3002\n\n\u51b2\u7a81 \u2192 409\u3002\u6279\u91cf\u5bfc\u5165 / \u521b\u5efa\u7528\u6237\u65f6\u7684\"\u8def\u5f84\u89e3\u6790\u81ea\u52a8\u5efa\" \u8d70 /resolve \u7aef\u70b9\u3002",
        "operationId": "create_department_api_v1_departments_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDepartmentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/meta": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Get Client Config",
        "description": "\u8fd4\u56de\u524d\u7aef\u6240\u9700\u7684\u540e\u7aef\u5e38\u91cf\uff08\u5355\u4e00\u771f\u76f8\u6e90\uff09\u3002\u503c\u9759\u6001\uff0c\u524d\u7aef\u53d6\u4e00\u6b21\u7f13\u5b58\u5373\u53ef\u3002",
        "operationId": "get_client_config_api_v1_meta_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientConfigResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/notifications": {
      "get": {
        "tags": [
          "notifications"
        ],
        "summary": "Get Notifications",
        "operationId": "get_notifications_api_v1_notifications_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteNotificationsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/skills": {
      "get": {
        "tags": [
          "skills"
        ],
        "summary": "List Skills",
        "description": "\u5217\u51fa\u5bf9\u5f53\u524d\u7528\u6237\u53ef\u89c1\u7684 skill + \u6709\u6548\u542f\u7528\u6001\u3002",
        "operationId": "list_skills_api_v1_skills_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminSkillListResponse": {
        "properties": {
          "skills": {
            "items": {
              "$ref": "#/components/schemas/AdminSkillItem"
            },
            "type": "array",
            "title": "Skills",
            "description": "All shared skills, including seeded read-only ones"
          }
        },
        "type": "object",
        "required": [
          "skills"
        ],
        "title": "AdminSkillListResponse",
        "description": "GET /api/v1/admin/skills response."
      },
...
      "Body_send_message_api_v1_chat_post": {
        "properties": {
          "payload": {
            "type": "string",
            "title": "Payload"
          },
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "payload",
          "files"
        ],
        "title": "Body_send_message_api_v1_chat_post"
      },
...
      "ChatResponse": {
        "properties": {
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id",
            "description": "Conversation ID"
          },
          "message_id": {
            "type": "string",
            "title": "Message Id",
            "description": "New message ID"
          },
          "stream_url": {
            "type": "string",
            "title": "Stream Url",
            "description": "SSE endpoint URL for streaming"
          }
        },
        "type": "object",
        "required": [
          "conversation_id",
          "message_id",
          "stream_url"
        ],
        "title": "ChatResponse",
        "description": "POST /api/v1/chat response"
      },
      "ClientConfigResponse": {
        "properties": {
          "compaction_token_threshold": {
            "type": "integer",
            "title": "Compaction Token Threshold",
            "description": "Effective token sum at which the lead model auto-compacts, calculated as its configured context_window minus the service reserve. Used by the frontend as the context-usage gauge denominator."
          },
          "lead_agent_model": {
            "type": "string",
            "title": "Lead Agent Model",
            "description": "Model identifier from the materialized DB lead_agent row (e.g. 'qwen3.7-max'). Turn execution reads that same registry snapshot; the composer badge therefore reflects the effective model rather than a per-worker agent MD cache."
          },
          "max_upload_size": {
            "type": "integer",
            "title": "Max Upload Size",
            "description": "Per-file upload byte limit (MAX_UPLOAD_SIZE). The composer uses it to pre-reject an oversize file with instant feedback instead of staging + POSTing it for a backend 422. Backend stays authoritative; the batch TOTAL is capped separately at the proxy layer (not surfaced here \u2014 it lives in Caddy config, outside src/config.py)."
          },
          "max_private_skills": {
            "type": "integer",
            "minimum": -1.0,
            "title": "Max Private Skills",
            "description": "Maximum number of private skills a user may own. -1 means unlimited, 0 disables personal skill imports, and a positive value is the limit. Shared/admin-imported skills do not count."
          },
          "message_feedback_max_detail_chars": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Message Feedback Max Detail Chars",
            "description": "Maximum optional detail length for one message feedback record."
          }
        },
        "type": "object",
        "required": [
          "compaction_token_threshold",
          "lead_agent_model",
          "max_upload_size",
          "max_private_skills",
          "message_feedback_max_detail_chars"
        ],
        "title": "ClientConfigResponse",
        "description": "GET /api/v1/meta response \u2014 runtime constants for the frontend."
      },
...
      "ConversationListResponse": {
        "properties": {
          "conversations": {
            "items": {
              "$ref": "#/components/schemas/ConversationSummary"
            },
            "type": "array",
            "title": "Conversations",
            "description": "Conversation list"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total count"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "Whether more results exist"
          }
        },
        "type": "object",
        "required": [
          "conversations",
          "total",
          "has_more"
        ],
        "title": "ConversationListResponse",
        "description": "GET /api/v1/chat response"
      },
...
      "CreateDepartmentRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "description": "Department name"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id",
            "description": "Parent department id; null = top-level"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateDepartmentRequest",
        "description": "POST /api/v1/departments request body"
      },
...
      "DepartmentListResponse": {
        "properties": {
          "departments": {
            "items": {
              "$ref": "#/components/schemas/DepartmentResponse"
            },
            "type": "array",
            "title": "Departments"
          }
        },
        "type": "object",
        "required": [
          "departments"
        ],
        "title": "DepartmentListResponse",
        "description": "GET /api/v1/departments response \u2014 \u540c\u7ea7\u5217\u8868"
      },
      "DepartmentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "user_count": {
            "type": "integer",
            "title": "User Count",
            "description": "\u8be5\u90e8\u95e8\u4e0b\u76f4\u5c5e\u7528\u6237\u6570\uff08\u4e0d\u542b\u5b50\u90e8\u95e8\uff09"
          },
          "child_count": {
            "type": "integer",
            "title": "Child Count",
            "description": "\u76f4\u63a5\u5b50\u90e8\u95e8\u6570"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "parent_id",
          "name",
          "user_count",
          "child_count",
          "created_at",
          "updated_at"
        ],
        "title": "DepartmentResponse",
        "description": "Single department response (\u542b\u76f4\u5c5e\u8ba1\u6570)"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError",
        "required": [
          "detail"
        ]
      },
...
      "SiteNotificationsResponse": {
        "properties": {
          "notifications": {
            "items": {
              "$ref": "#/components/schemas/SiteNotification"
            },
            "type": "array",
            "title": "Notifications"
          },
          "revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Revision"
          }
        },
        "type": "object",
        "required": [
          "notifications",
          "revision"
        ],
        "title": "SiteNotificationsResponse"
      },
...
      "SkillListResponse": {
        "properties": {
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillItem"
            },
            "type": "array",
            "title": "Skills",
            "description": "Skills visible to the user, with effective enabled state."
          }
        },
        "type": "object",
        "required": [
          "skills"
        ],
        "title": "SkillListResponse",
        "description": "GET /api/v1/skills response."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type",
          "input",
          "ctx"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}